reorganizing how Meta interacts with AWSMeta
